split file for lyapinf toy example
